Split large WasmDriver into submodules
