+++ src/proto/mod.rs
@@ -1,6 +1,16 @@
+pub mod google {
+    pub mod protobuf {
+        include!("google.protobuf.rs");
+    }
+}
 pub mod hashicorp {
     pub mod nomad {
         pub mod plugins {
+            pub mod base {
+                pub mod proto {
+                    include!("hashicorp.nomad.plugins.base.proto.rs");
+                }
+            }
             pub mod shared {
                 pub mod structs {
                     include!("hashicorp.nomad.plugins.shared.structs.rs");
@@ -14,16 +24,6 @@ pub mod hashicorp {
                     include!("hashicorp.nomad.plugins.drivers.proto.rs");
                 }
             }
-            pub mod base {
-                pub mod proto {
-                    include!("hashicorp.nomad.plugins.base.proto.rs");
-                }
-            }
         }
-    }
-}
-pub mod google {
-    pub mod protobuf {
-        include!("google.protobuf.rs");
     }
 }
